pyproject.toml
[project]
name = "mcp-server-gmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.4",
    "google-api-python-client>=2.194.0",
    "google-auth-oauthlib>=1.3.1",
    "httpx>=0.28.1",
]
